Add special handling for '-Cmd' suffixed PVs
The Sirius naming system convention assumes that PVs suffixed with '-Cmd'
cause an IOC action to happen whenever they are processed, regardless
of the value the PV was set to.

In order to provide support for this behavior and keep it compatible
with the etherip module periodic scanning behavior, a new string
template was added to the auto-gen scripts, called 'bo_cmd_template'.

This template creates a bo record which sends a 1 to the PLC (by means
of 2 additional auxiliary records) whenever it is processed, regardless
of the value sent to it by CA.