scripts/: Remove test_ioc_gen.sh
Remove this script since the 'scripts/Makefile' can also be used as an example
on how to use the db auto-gen script for the Delta application, making
the test_ioc_gen script.sh redundant.

On the other hand, the test_ioc_gen_multi.sh script is kept since the
ioc_gen_delta_multi.py script is not yet used in the Makefile, but could be
useful in the future.