delta: use etherip scan flag for rec output links
Problem:

During Delta IOC tests it was verified that specifying a scan period
at the SCAN field of an output record would result in it sometimes
failing to send a new value to the PLC.

In the tests performed, when the SCAN field was set to "Passive" and
a 'S' scan flag was used in the OUT field, the record would not
fail sending values when processed anymore.

Solution:

The template used for generating the Delta IOC was modified to use
scan flags for output records instead of specifying the SCAN field
different than Passive.