sncSabia: Change monitored cmd PVs variable names
Change variable names associated to command PVs that are
monitored by sncSabia sequencer program, in order to make
it clear that these are not variables that the sequencer
can use to send commands. They are instead monitored to
verify when user commands were sent to the sequencer.

Change: suffix "_cmd" to "_evt", when the cmd PV is intendend
to be read only.