sncSabia.stt: Check if scan is necessary before attempt
Change sncSabia state machine, so that after receiving a scan command,
the need for motion during the scan is checked before attempting the scan.

If the scan distance is smaller than the position tolerance value,
the scan is ignored.

Before, the scan would be ignored but only after the axes were
positioned at the scan start position. Now, the axes do not
move to the scan start position if the scan is going to be
ignored later.