sncSabia.stt: Simplify polarization change procedure
The previous procedure of changing the undulator polarization
consisted of moving the undulator to the minimum gain configuration
then adjusting the polarization and finally recovering the initial
gain configuration.

Now, after changing the undulator polarization, the device is left
at the minimum gain position, instead of recovering the previous
configuration.

This change was required by the users.